Delete method functioning, reorganised some helper function
