+++ DefaultConfig.conf
@@ -27,7 +27,7 @@ server {
 
 server {
    listen `host:8888`;
-         server_name  `2test_name`;
+        server_name  `2test_name`;
    error_page  504  /504.html;
    error_page   500 502  /50x.html;
    client_max_body_size 255;
@@ -35,6 +35,7 @@ server {
    location / {
       allow GET;
       allow POST;
+	  allow DELETE;
       redirection /newer_path;
       root   /usr/share/nginx/html;
       listing on/off;
@@ -43,7 +44,9 @@ server {
       CGI stuff?
    }
    location /example_folder {
+      allow GET;
       allow POST;
+	  allow DELETE;
       redirection /new_example_folder;
       root   /usr/share/nginx/html;
       listing on/off;
