+++ src/WebServ/constants.hpp
@@ -3,7 +3,7 @@
 
 // Status codes
 # define	ERRORS			400
-
+# define 	REQUEST_TIMEOUT 408
 # define	OK 				200
 # define	CREATED			201
 # define	UNAUTHORIZED	401
@@ -12,6 +12,7 @@
 # define	INT_ERROR		500
 # define	ACCEPTED		202
 # define	NO_CONTENT		204
+
 // Response formatting lines
 # define	HTTP_HEADER	"HTTP/1.1 "
 # define 	END_HEADER	"\r\n\r\n"
