+++ test/index.html
@@ -1,11 +1,11 @@
 <!DOCTYPE html>
+<html lang="en">
 <head>
-    <meta charset="UTF-8" />
-    <link rel="stylesheet" type="text/css" href="index.css" /> 
+    <meta charset="UTF-8">
+    <link rel="stylesheet" type="text/css" href="index.css"> 
     <title>Link CSS to HTML</title>
 </head>
-<html>
-	<body>
-		<h1>Hello world!</h1>
-	</body>
+<body>
+    <h1>Hello world!</h1>
+</body>
 </html>
